CE support with an ugly hack to select/signaler to make them faster.
The Windows CE support - for CE >= 4.2 - should work more or less out of the
box. Several functions are missing from WinSock as implemented in CE 4.2, so
workarounds are needed.

Sockets to localhost are very slow under CE, because they seem to be going
through the entire network stack all the way to the adapter driver. This means
a send() takes up to 2ms on my test hardware, which in turns makes select() and
the signaler unacceptably slow - a zmq_send() of a seven-part message would
take up to 40 milliseconds.

The hack around this problem is to modify the signaller to use WinSock events
instead of sockets. These are much faster, but this means the signaller cannot
provide FDs to pass to select(). Hacking around this means implementing a custom
select() function, winselect(), which accepts WinSock FDs as well as WinSock events
put in the fd_set as if they were real FDs. The function can differentiate.